Add Organizer Events Button

+++ Inlämning_Asp/Pages/Delete_event.cshtml.cs
@@ -1,24 +1,24 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Inlämning_Asp.Data;
+using Inlämning_Asp.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
-using Inlämning_Asp.Data;
-using Inlämning_Asp.Models;
+using System.Threading.Tasks;
 
 namespace Inlämning_Asp.Pages
 {
+    [Authorize(Roles = "Organizer")]
     public class Delete_eventModel : PageModel
     {
-        private readonly Inlämning_Asp.Data.EventDbContext _context;
+        private readonly EventDbContext _context;
 
-        public Delete_eventModel(Inlämning_Asp.Data.EventDbContext context)
+        public Delete_eventModel(EventDbContext context)
         {
             _context = context;
         }
 
+        [BindProperty]
         public Event Event { get; set; }
 
         public async Task<IActionResult> OnGetAsync(int? id)
@@ -36,42 +36,21 @@ namespace Inlämning_Asp.Pages
             }
             return Page();
         }
-        //public async Task<IActionResult> OnPostAsync(int? id)
-        //{
-        //    if (id == null)
-        //    {
-        //        return NotFound();
-        //    }
-
-        //    Event = await _context.Events.Include(e => e.Events).FirstOrDefaultAsync(m => m.Id == id);
-
-        //    if (Event == null)
-        //    {
-        //        return NotFound();
-        //    }
-
-        //    var events = await _context.Events.FirstOrDefaultAsync();
-
-        //    if (!Event.Events.Contains(events))
-        //    {
-        //        Event.Events.Remove(events);
-        //        await _context.SaveChangesAsync();
-        //    }
-
-
-
-        //return Page();
-
 
         public async Task<IActionResult> OnPostAsync(int? id)
         {
             if (id == null)
             {
-                return Page();
+                return NotFound();
             }
 
-            _context.Events.Remove(Event);
-            await _context.SaveChangesAsync();
+            Event = await _context.Events.FindAsync(id);
+
+            if (Event != null)
+            {
+                _context.Events.Remove(Event);
+                await _context.SaveChangesAsync();
+            }
 
             return RedirectToPage("./Index");
         }
